rename and bugfix
