+++ match.go
@@ -4,30 +4,19 @@ import (
 	"errors"
 	"sort"
 
-	"github.com/cearius/go-charmagic/pkg/m/multibyte"
-	"github.com/cearius/go-charmagic/pkg/m/singlebyte"
-	"github.com/cearius/go-charmagic/pkg/m/unicode"
+	"github.com/cearius/go-charmagic/pkg/m"
 	"github.com/cearius/go-charmagic/pkg/matching"
-	"github.com/cearius/go-charmagic/pkg/util"
 )
 
 var ErrNoMatchFound = errors.New("no match found")
-
-func CreateAllMatchers() []matching.Matcher {
-	return util.Collect(
-		unicode.Create_Unicode_Matchers(),
-		singlebyte.Create_SingleByte_Matchers(),
-		multibyte.Create_MultiByte_Matchers(),
-	)
-}
 
 // MatchAll matches all supported encodings against the input bytes and assigns a confidence score.
 // The higher the score the more likely the algorithm deemed the chance of having picked the right encoding.
 func MatchAll(buf []byte) (results matching.Results) {
 	input := matching.FromBytes(buf)
 
-	for _, m := range CreateAllMatchers() {
-		results = append(results, m.Match(input))
+	for _, matcher := range m.CreateAllMatchers() {
+		results = append(results, matcher.Match(input))
 	}
 
 	sort.Stable(results)
